Add SwiftUI project template
