+++ Templates/SwiftUI App.xctemplate/App.swift
@@ -3,15 +3,15 @@
 import SwiftUI
 
 @main
-struct ___VARIABLE_appIdentifier___App: App {
+struct ___PACKAGENAME:identifier___App: App {
     @UIApplicationDelegateAdaptor
-    private var appDelegate: ___VARIABLE_appIdentifier___AppDelegate
+    private var appDelegate: ___PACKAGENAME:identifier___AppDelegate
 
     @StateObject
-    private var coordinator: ___VARIABLE_appIdentifier___AppCoordinator
+    private var coordinator: ___PACKAGENAME:identifier___AppCoordinator
 
     init() {
-        let coordinator = ___VARIABLE_appIdentifier___AppCoordinator(container: Container())
+        let coordinator = ___PACKAGENAME:identifier___AppCoordinator(container: ___PACKAGENAME:identifier___Container())
         self._coordinator = StateObject(wrappedValue: coordinator)
         self.appDelegate.delegate = coordinator
     }
