--- Templates/SwiftUI App.xctemplate/___VARIABLE_appIdentifier___Container.swift
@@ -1,16 +1,0 @@
-//  ___FILEHEADER___
-
-import FuturedArchitecture
-
-final class ___VARIABLE_appIdentifier___Container {
-    private(set) var dataCache = DataCache(value: ___VARIABLE_appIdentifier___CacheModel())
-
-    init() {
-        // Init services
-    }
-
-    func resetContainer() {
-        // Reset services
-        self.dataCache = DataCache(value: ___VARIABLE_appIdentifier___CacheModel())
-    }
-}
